Rangement du dossier Model, commencement du visuel de la page login
