+++ public/src/Controllers/Comment/EditComment.php
@@ -2,7 +2,7 @@
 
 namespace Application\Controllers\Comment\Update;
 
-require_once('src/Model/comment.php');
+require_once('src/Model/Comment.php');
 require_once('src/Lib/DatabaseConnection.php');
 
 use Application\Model\Comment\CommentRepository;
