+++ public/src/Model/Repository/PostRepository.php
@@ -1,17 +1,9 @@
 <?php
 
-namespace Application\Model\Post;
-
-require_once ('src/Lib/DatabaseConnection.php');
+namespace Application\Model\Repository\Post;
 
 use Application\lib\Database\DatabaseConnection;
-
-class Post {
-    public string $title;
-    public string $content;
-    public string $frenchCreationDate;
-    public string $identifier;
-}
+use Application\Model\Post\Post;
 
 class PostRepository {
     public DatabaseConnection $connection;
